refactor(desktop): move to formally express dependency
