+++ harper-desktop/src-tauri/src/highlighter_service/highlighter_process.rs
